pipeline: use size_t for buffer offset and size
Both OpenGL and Vulkan backend expect size-based types.

+++ libnopegl/src/backends/gl/pipeline_gl.c
@@ -323,8 +323,8 @@ static void set_buffers(struct pipeline *s, struct glcontext *gl)
         const struct buffer *buffer = buffer_binding->buffer;
         const struct buffer_gl *buffer_gl = (const struct buffer_gl *)buffer;
         const struct pipeline_buffer_desc *buffer_desc = &buffer_binding->desc;
-        const int offset = buffer_desc->offset;
-        const int size = buffer_desc->size ? buffer_desc->size : buffer->size;
+        const size_t offset = buffer_desc->offset;
+        const size_t size = buffer_desc->size ? buffer_desc->size : buffer->size;
         ngli_glBindBufferRange(gl, buffer_binding->type, buffer_desc->binding, buffer_gl->id, offset, size);
     }
 }
@@ -709,7 +709,7 @@ int ngli_pipeline_gl_update_texture(struct pipeline *s, int index, const struct 
     return 0;
 }
 
-int ngli_pipeline_gl_update_buffer(struct pipeline *s, int index, const struct buffer *buffer, int offset, int size)
+int ngli_pipeline_gl_update_buffer(struct pipeline *s, int index, const struct buffer *buffer, size_t offset, size_t size)
 {
     struct pipeline_gl *s_priv = (struct pipeline_gl *)s;
