+++ libnopegl/src/backends/vk/pipeline_vk.h
@@ -55,7 +55,7 @@ int ngli_pipeline_vk_set_resources(struct pipeline *s, const struct pipeline_res
 int ngli_pipeline_vk_update_attribute(struct pipeline *s, int index, const struct buffer *buffer);
 int ngli_pipeline_vk_update_uniform(struct pipeline *s, int index, const void *value);
 int ngli_pipeline_vk_update_texture(struct pipeline *s, int index, const struct texture *texture);
-int ngli_pipeline_vk_update_buffer(struct pipeline *s, int index, const struct buffer *buffer, int offset, int size);
+int ngli_pipeline_vk_update_buffer(struct pipeline *s, int index, const struct buffer *buffer, size_t offset, size_t size);
 void ngli_pipeline_vk_draw(struct pipeline *s, int nb_vertices, int nb_instances);
 void ngli_pipeline_vk_draw_indexed(struct pipeline *s, const struct buffer *indices, int indices_format, int nb_vertices, int nb_instances);
 void ngli_pipeline_vk_dispatch(struct pipeline *s, int nb_group_x, int nb_group_y, int nb_group_z);
