+++ libnopegl/src/pipeline.h
@@ -22,6 +22,8 @@
 #ifndef PIPELINE_H
 #define PIPELINE_H
 
+#include <stdlib.h>
+
 #include "buffer.h"
 #include "darray.h"
 #include "graphicstate.h"
@@ -62,8 +64,8 @@ struct pipeline_buffer_desc {
     int binding;
     int access;
     int stage;
-    int offset;
-    int size;
+    size_t offset;
+    size_t size;
 };
 
 struct pipeline_attribute_desc {
@@ -129,7 +131,7 @@ int ngli_pipeline_set_resources(struct pipeline *s, const struct pipeline_resour
 int ngli_pipeline_update_attribute(struct pipeline *s, int index, const struct buffer *buffer);
 int ngli_pipeline_update_uniform(struct pipeline *s, int index, const void *value);
 int ngli_pipeline_update_texture(struct pipeline *s, int index, const struct texture *texture);
-int ngli_pipeline_update_buffer(struct pipeline *s, int index, const struct buffer *buffer, int offset, int size);
+int ngli_pipeline_update_buffer(struct pipeline *s, int index, const struct buffer *buffer, size_t offset, size_t size);
 void ngli_pipeline_draw(struct pipeline *s, int nb_vertices, int nb_instances);
 void ngli_pipeline_draw_indexed(struct pipeline *s, const struct buffer *indices, int indices_format, int nb_indices, int nb_instances);
 void ngli_pipeline_dispatch(struct pipeline *s, int nb_group_x, int nb_group_y, int nb_group_z);
